+++ packages/keycloak-user-management/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.18](https://github.com/OpenSRP/web/compare/@opensrp/user-management@0.1.16...@opensrp/user-management@0.1.18) (2021-06-30)
+
+**Note:** Version bump only for package @opensrp/user-management
+
+## [0.1.17](https://github.com/OpenSRP/web/compare/@opensrp/user-management@0.1.16...@opensrp/user-management@0.1.17) (2021-06-21)
+
+**Note:** Version bump only for package @opensrp/user-management
+
+## [0.1.16](https://github.com/OpenSRP/web/compare/@opensrp/user-management@0.1.15...@opensrp/user-management@0.1.16) (2021-06-17)
+
+**Note:** Version bump only for package @opensrp/user-management
+
 ## 0.1.13 (2021-02-24)
 
 **Note:** Version bump only for package @opensrp/user-management